refactor(activerecord): extract CollectionProxy to associations/collection-proxy.ts
Move CollectionProxy (~865 lines) from associations.ts to its
Rails-canonical location at associations/collection-proxy.ts.

Export resolveModel, fireAssocCallbacks, buildHasManyRelation from
associations.ts so CollectionProxy can import them.

All imports updated to reference new location directly.
api:compare: 48/597 (8%), 14 misplaced (was 15).
